1660 - move ProductionStat from UI to stats package
