+++ game-core/src/main/java/games/strategy/triplea/ui/StatPanel.java
@@ -9,8 +9,8 @@ import games.strategy.engine.data.UnitType;
 import games.strategy.engine.data.events.GameDataChangeListener;
 import games.strategy.engine.stats.AbstractStat;
 import games.strategy.engine.stats.IStat;
+import games.strategy.engine.stats.ProductionStat;
 import games.strategy.triplea.Constants;
-import games.strategy.triplea.Properties;
 import games.strategy.triplea.attachments.PlayerAttachment;
 import games.strategy.triplea.attachments.TerritoryAttachment;
 import games.strategy.triplea.delegate.Matches;
@@ -409,28 +409,6 @@ class StatPanel extends AbstractStatPanel {
       gameData = data;
       gameData.addDataChangeListener(this);
       isDirty = true;
-    }
-  }
-
-  static class ProductionStat extends AbstractStat {
-    @Override
-    public String getName() {
-      return "Production";
-    }
-
-    @Override
-    public double getValue(final GamePlayer player, final GameData data) {
-      final int production =
-          data.getMap().getTerritories().stream()
-              .filter(place -> place.getOwner().equals(player))
-              .filter(Matches.territoryCanCollectIncomeFrom(player, data))
-              .mapToInt(TerritoryAttachment::getProduction)
-              .sum();
-      /*
-       * Match will Check if terr is a Land Convoy Route and check ownership
-       * of neighboring Sea Zone, or if contested
-       */
-      return (double) production * Properties.getPuMultiplier(data);
     }
   }
